extract remove-nodes from prune-newick.js
