+++ server/ent/ent.js
@@ -3,7 +3,7 @@
 const combs = require('combinations-generator')
 const uniqBy = require('lodash/uniqBy')
 const remove = require('lodash/remove')
-const { removeNodes } = require('../lib/prune-newick')
+const { removeNodes } = require('../lib/remove-nodes')
 
 const ENABLE_DEBUG = false
 let debug = ENABLE_DEBUG ? console.log.bind(console) : () => {}
